requirements.txt
Flask
flask_login
flask_session
cryptography
mysql.connector
cron_descriptor
crontab
uuid
datetime
flask_bcrypt
werkzeug
werkzeug
logging
